requirements.txt
discord.py~=1.4.1
PyYAML~=3.13
